+++ core/core-api/src/main/java/com/ticket/api/controller/v1/MemberController.java
@@ -1,0 +1,31 @@
+package com.ticket.api.controller.v1;
+
+import com.ticket.api.controller.v1.request.AddMemberRequest;
+import com.ticket.api.controller.v1.response.MemberResponse;
+import com.ticket.domain.member.Member;
+import com.ticket.domain.member.MemberService;
+import com.ticket.support.response.ApiResponse;
+import jakarta.validation.Valid;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/api/v1/member")
+public class MemberController {
+
+    private final MemberService memberService;
+
+    public MemberController(final MemberService memberService) {
+        this.memberService = memberService;
+    }
+
+    @PostMapping
+    public ApiResponse<Long> register(@RequestBody @Valid AddMemberRequest request) {
+        return ApiResponse.success(memberService.register(request.toAddMember()));
+    }
+
+    @GetMapping("/{memberId}")
+    public ApiResponse<MemberResponse> findById(@PathVariable Long memberId) {
+        final Member findMember = memberService.findById(memberId);
+        return ApiResponse.success(new MemberResponse(findMember.getId(), findMember.getEmail().getEmail(), findMember.getName()));
+    }
+}
