--- core/core-api/src/main/java/com/ticket/member/controller/MemberController.java
@@ -1,26 +1,0 @@
-package com.ticket.member.controller;
-
-import com.ticket.member.dto.AddMemberRequest;
-import com.ticket.domain.member.MemberService;
-import com.ticket.support.response.ApiResponse;
-import jakarta.validation.Valid;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/member")
-public class MemberController {
-
-    private final MemberService memberService;
-
-    public MemberController(final MemberService memberService) {
-        this.memberService = memberService;
-    }
-
-    @PostMapping
-    public ApiResponse<Long> register(@RequestBody @Valid AddMemberRequest request) {
-        return ApiResponse.success(memberService.register(request.toAddMember()));
-    }
-}
